+++ packages/perseus/src/prompt-types.ts
@@ -16,7 +16,7 @@ import type {NumberLinePromptJSON} from "./widget-ai-utils/number-line/number-li
 import type {NumericInputPromptJSON} from "./widget-ai-utils/numeric-input/prompt-utils";
 import type {OrdererPromptJSON} from "./widget-ai-utils/orderer/orderer-ai-utils";
 import type {PassagePromptJSON} from "./widget-ai-utils/passage/passage-ai-utils";
-import type {PassageRefPromptJSON} from "./widget-ai-utils/passage-ref/prompt-utils";
+import type {PassageRefPromptJSON} from "./widget-ai-utils/passage-ref/passage-ref-ai-utils";
 import type {RadioPromptJSON} from "./widget-ai-utils/radio/prompt-utils";
 import type {SorterPromptJSON} from "./widget-ai-utils/sorter/prompt-utils";
 import type {UnsupportedWidgetPromptJSON} from "./widget-ai-utils/unsupported-widget";
