+++ packages/perseus/src/widget-ai-utils/phet-simulation/phet-simulation-ai-utils.ts
